+++ NEWS.md
@@ -1,4 +1,7 @@
 # climateData 2.2.3
+
+- use package prefix `reproducible::` for `linkOrCopy`
+
 - removed `calcATA` and `calcCMI` as pre-existing functions, as the functions were incompatible with 
 use cases of calcAsIS containing multiple normal periods
 - fixed errors in `calcAsIs` where multiple climate normal periods and variables are retrieved